==> Feature Engineer/2_data_for_predict.ipynb <==
import pandas as pd


df = pd.read_csv('aligned_window_based_engineered_features.csv')


# Shift the 'y' column by -1 to align y_i+1 with x_i
df['next_16day_ndvi_mean'] = df['ndvi_mean'].shift(-1)

# Drop the last row since it will have a NaN value in the 'y_shifted' column
df = df.dropna(subset=['next_16day_ndvi_mean'])


df.drop(columns=['ndvi_mean'], inplace=True)


df.to_csv('prediction_aligned_window_based_engineered_features.csv', index=False)
